requirements.txt
requests>=2.31,<3
pandas>=2.0,<3
openpyxl>=3.1,<4
beautifulsoup4>=4.12,<5
python-dotenv>=1.0,<2
google-genai>=1.0,<2
ddgs>=9.0,<10
streamlit>=1.30,<2
pydantic>=2.0,<3
tenacity>=8.2,<10